paillier and log fix
1. pailler modify to paillier;

2. dfs init waiting log print only once
