+++ libpaillier/CMakeLists.txt
@@ -1,0 +1,7 @@
+aux_source_directory(. SRC_LIST)
+
+file(GLOB HEADERS "*.h")
+
+add_library(pailler ${SRC_LIST} ${HEADERS})
+
+install(TARGETS pailler RUNTIME DESTINATION bin ARCHIVE DESTINATION lib LIBRARY DESTINATION lib)
