+++ libpaillier/typedef.h
@@ -1,0 +1,27 @@
+#ifndef __HEADER_TYPEDEF_H__
+#define __HEADER_TYPEDEF_H__
+
+#include <stdint.h>
+#include "macro.h"
+
+#ifdef __cplusplus 
+extern "C"{
+#endif
+
+	typedef unsigned char		U8;
+	typedef unsigned short		U16;
+	typedef uint32_t		    U32;
+	typedef uint64_t	        U64;
+
+	typedef char				S8;
+	typedef short				S16;
+	typedef int32_t				S32;
+	typedef int64_t 			S64;	
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
+
+
