--- tool/java/Pailler/.classpath
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="lib" path="lib/bcprov-jdk15on-1.54.jar"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry kind="lib" path="lib/ezmorph-1.0.3.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
